refactor: newlayout file structure, component file namings
