+++ src/newlayout/components/Map/Controls/Control.tsx
