+++ src/newlayout/components/Map/Dynamic.tsx
