+++ src/newlayout/components/UserGuide/UserGuide.tsx
